requirements.txt
langchain-community
langchain
openai
huggingface_hub
python-dotenv
streamlit
openai